M8b: Orca Whirlpool direct-pool adapter; all five emergency-route families supported
Decodes the Whirlpool, fixed and dynamic tick arrays and the oracle account
(adaptive fee tiers); replays the program's swap loop with its tick math,
compute_swap, tick crossings and adaptive fee manager; builds swap_v2 with the
three tick arrays the program requires plus the oracle PDA, carrying an absent
array as its PDA exactly as the program does. Validated live on three
whirlpools with the local quote equal to the simulated output. getBalance no
longer rejects treasuries above 2^53 lamports.

+++ libs/solana-hard-state/src/rpc/client.ts
@@ -92,7 +92,8 @@ export const TokenLargestAccounts = z.object({
   value: z.array(TokenAmount.extend({ address: z.string() })).max(20),
 });
 
-export const Balance = z.object({ context: z.object({ slot: z.number().int().nonnegative() }), value: z.number().int().nonnegative() });
+/** Lamports arrive as a JSON number; a treasury above 2^53 lamports loses precision but must not fail to parse. */
+export const Balance = z.object({ context: z.object({ slot: z.number().int().nonnegative() }), value: z.number().nonnegative() });
 
 /** jsonParsed token account as returned by getTokenAccountsByOwner. */
 export const ParsedTokenAccount = z.object({
